@@ -1,10 +1,10 @@
 package com.pranayharjai7.myemotions;
 
-import androidx.appcompat.app.AppCompatActivity;
-
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.pranayharjai7.myemotions.LoginAndRegister.LoginActivity;
 import com.pranayharjai7.myemotions.databinding.ActivityLauncherBinding;
@@ -26,7 +26,7 @@ public class LauncherActivity extends AppCompatActivity {
             finish();
         };
 
-        binding.LauncherImage.postDelayed(runnable, 1000);
+        binding.LauncherImage.postDelayed(runnable, 500);
     }
 
     public void launcherConstraintLayoutClicked(View view) {
